- The "Scene::onCollision()" and "Scene::onEndCollision()" (mutial object collision callbacks) only get called when both objects involved have their collision callback active which hasn't changed however now the "self" collision callbacks on both SceneObject and their behaviors are called when the respective SceneObject has its collision callback active i.e. it does not require the other objects (the object it's colliding with) to be active.
An example of this would be not having to turn-on collision callbacks on static scenary to get a collision callback for a player.  In this case you could use the "onSelfCollision()" or "onSelfEndCollision()" callbacks but not the mutual scene callbacks if the other object(s) didn't have their collision callback turned-on.