Commit regarding this comment in SkeletonAsset.h: // BOZO - How to expose scale to scripts?. I have added the necessary code to take a variable defined in source (like mScale), and get it hook up to non-source systems (like TorqueScript or XML).
This work consists of three main steps:

1. Public setter and getter
2. Protected setter, getter, and writer check
3. The addProtectedField function

The public setter/getter can be found in SkeletonAsset.h, lines 52 and 53. The protected versions are further down on lines 81-83. Any variables you want to expose outsid of the source should be implemented in a class's initPersistFields function. You can see this in action in SkeletonAsset.cc, line 134. This is the usage for addProtectedField:

addProtectedField(Variable Name In Script, Variable Type, Size offset of variable in the class, the set function, the get function, the write function, the documentation string);

There are other macros for exposing a variable to script, each with its own reason for existing. We can discuss this more via Skype.