Using doxygen "groups" or "modules" as a major factor in keeping the documentation (reference material and all) with the source. - specifically, check the TAML Reference module and the Module Manager module for the engine.
To do this:
First the reference material from github wiki is brought over as a *.dox file.  doxygen understands markdown
95% well so not many tweaks are required.  By usingn a .dox file, the large reference is kept together and not
injected within a specific class.  Using the .dox is fairly standard practice.

The reference doc is placed in the dir that most closely represents it, such as module/module.dox or persistence/taml/taml.dox.

Classes in the module add themselves to their respective module with "/// @ingroup TAMLGroup" in the class header.

Now the markdown reference is inserted as a "module" and has references to all classes that are involved.

note: because the "modules" page in the output had lots of little groups, (such as dglThis and dglThat), I've also been pushing these groups into super groups to keep it organized (such as dglGroup which now wraps dglThis and dglThat).