Replaced all instances of onTouchMoved to onTouchDragged in the Sandbox modules.
This happened in AngleToy, MoveToToy, PickingToy, and RotateToToy. onTouchMoved does not get called on iOS or Android. It's purely a mouse function, despite the name. By switching to onTouchDragged, the toys should now work on mobile platforms. This does have a side effect on desktop platforms - users now need to mouse click down or click/drag in these toys instead of just moving the mouse.