Fixed Bug in XML
SimXMLDocument would crash if you tried to read a field that had a blank
value.  This happened because it checks the text of the first child
element when there is no first child element.  This fixes the problem by
checking to see if the first child element is null first.