Added Fading Functionality
FadeTo allows any SceneObject to dynamically change its BlendColor over
a period of time!  It includes support functions like cancelFadeTo()
that were modeled after moveTo() and rotateTo().  When the fade is
complete an onFadeToComplete() callback is called.  This has all been
thoroughly tested and is ready to move to other dev environments.