Tab Book Update
Previously the tab book was barely functional. Positioning the tabs to the right, left, or bottom did not work and there was no option to have the tabs draw without images. These changes fix that, but leave off support (for now) of image skinned tabs. It makes heavy use of the new border model.