Initial Editor Setup
There are no editors yet, but this does set up the folder and file structures needed to make the editors. It moves the toys out of a "modules" folder and puts them into a "toybox" folder. It also creates an editor folder which will house the editors and a library folder which will hold reusable modules.