Changes to guiArrayCtrl
The array control was useless by itself and served as a parent control for a number of other controls. It was removed as a console object and was made a purely abstract class that requires renderCell to be overridden. The headers for columns and rows were never used and were removed. Events were renamed from Mouse to Touch.